# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /diet/profile:
    get:
      tags: [diet]
      operationId: getDietProfile
      summary: Get user diet profile
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: User diet profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DietProfileEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      tags: [diet]
      operationId: saveDietProfile
      summary: Save or update user diet profile
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DietProfileInput'
      responses:
        '200':
          description: Profile saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DietProfileEnvelope'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /diet/plan:
    get:
      tags: [diet]
      operationId: getDietPlan
      summary: Get the user personalized diet plan
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Diet plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DietPlanEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Profile not set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DietProfileInput:
      type: object
      required: [age, gender, weightKg, heightCm, activityLevel, goal]
      properties:
        age:
          type: integer
          minimum: 1
          maximum: 120
        gender:
          type: string
          enum: [male, female]
        weightKg:
          type: number
          minimum: 20
          maximum: 500
        heightCm:
          type: number
          minimum: 50
          maximum: 300
        activityLevel:
          type: string
          enum: [sedentary, light, moderate, active, very_active]
        goal:
          type: string
          enum: [lose, maintain, gain]
        restrictions:
          type: [string, 'null']
    DietProfile:
      type: object
      required: [id, userId, age, gender, weightKg, heightCm, activityLevel, goal, tmb, tdee, targetCalories]
      properties:
        id:
          type: string
        userId:
          type: string
        age:
          type: integer
        gender:
          type: string
        weightKg:
          type: number
        heightCm:
          type: number
        activityLevel:
          type: string
        goal:
          type: string
        restrictions:
          type: [string, 'null']
        tmb:
          type: number
        tdee:
          type: number
        targetCalories:
          type: number
    DietProfileEnvelope:
      type: object
      required: [profile]
      properties:
        profile:
          oneOf:
            - $ref: '#/components/schemas/DietProfile'
            - type: 'null'
    Meal:
      type: object
      required: [name, time, calories, protein, carbs, fat, foods]
      properties:
        name:
          type: string
        time:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        foods:
          type: array
          items:
            $ref: '#/components/schemas/Food'
    Food:
      type: object
      required: [name, quantity, calories, protein, carbs, fat]
      properties:
        name:
          type: string
        quantity:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
    DietPlan:
      type: object
      required: [totalCalories, totalProtein, totalCarbs, totalFat, meals]
      properties:
        totalCalories:
          type: number
        totalProtein:
          type: number
        totalCarbs:
          type: number
        totalFat:
          type: number
        meals:
          type: array
          items:
            $ref: '#/components/schemas/Meal'
    DietPlanEnvelope:
      type: object
      required: [plan, profile]
      properties:
        plan:
          $ref: '#/components/schemas/DietPlan'
        profile:
          $ref: '#/components/schemas/DietProfile'