# lib/api-zod/src/generated/types/dietPlanEnvelope.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DietPlan } from "./dietPlan";
import type { DietProfile } from "./dietProfile";

export interface DietPlanEnvelope {
  plan: DietPlan;
  profile: DietProfile;
}