# lib/api-zod/src/generated/types/dietProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DietProfile {
  id: string;
  userId: string;
  age: number;
  gender: string;
  weightKg: number;
  heightCm: number;
  activityLevel: string;
  goal: string;
  /** @nullable */
  restrictions?: string | null;
  tmb: number;
  tdee: number;
  targetCalories: number;
}