# lib/api-zod/src/generated/types/dietProfileInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DietProfileInputActivityLevel } from "./dietProfileInputActivityLevel";
import type { DietProfileInputGender } from "./dietProfileInputGender";
import type { DietProfileInputGoal } from "./dietProfileInputGoal";

export interface DietProfileInput {
  /**
   * @minimum 1
   * @maximum 120
   */
  age: number;
  gender: DietProfileInputGender;
  /**
   * @minimum 20
   * @maximum 500
   */
  weightKg: number;
  /**
   * @minimum 50
   * @maximum 300
   */
  heightCm: number;
  activityLevel: DietProfileInputActivityLevel;
  goal: DietProfileInputGoal;
  /** @nullable */
  restrictions?: string | null;
}